requirements.txt
streamlit>=1.22
boto3
opensearch-py
requests-aws4auth
pdfminer.six
toml
